% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{drawHeatmap
for heatmapPathway app}
\usage{
drawHeatmap(
  input,
  z_data_mx,
  split_group,
  dge.p,
  DEGp_colFun,
  dge.logFC,
  logFC_colFun,
  dge.logFC.breaks,
  gseaRes,
  genes
)
}
\arguments{
\item{input}{shiny inpuy}

\item{z_data_mx}{z-scored data matrix}

\item{split_group}{global variable split_group}

\item{dge.p}{global variable}

\item{DEGp_colFun}{global variable}

\item{dge.logFC}{global variable}

\item{logFC_colFun}{global variable}

\item{dge.logFC.breaks}{global variable}

\item{gseaRes}{global variable gseaRes}

\item{genes}{global variable}
}
\description{
drawHeatmap
for heatmapPathway app
}
