% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getGenes}
\alias{getGenes}
\title{getGenes
for heatmapPathway app}
\usage{
getGenes(pthwy, input)
}
\arguments{
\item{pthwy}{global variable pthwy (res of getPthwy)}

\item{input}{}
}
\description{
getGenes
for heatmapPathway app
}
