% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getGseaRes}
\alias{getGseaRes}
\title{getGseaRes
for heatmapPathway app}
\usage{
getGseaRes(input)
}
\arguments{
\item{input}{shiny input}
}
\value{
GseaRes
}
\description{
getGseaRes
for heatmapPathway app
}
