% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getPCorMxNetwork}
\alias{getPCorMxNetwork}
\title{getPCorMxNetwork}
\usage{
getPCorMxNetwork(
  data_mx,
  plotIg = F,
  method = c("ppcor", "corpcor"),
  seed = 123,
  pth = 0.05
)
}
\arguments{
\item{data_mx}{input data matrix}

\item{plotIg}{whether or not to plot igraph network}

\item{method}{method to calculate partial correlation matrix}

\item{seed}{seed for igraph layout}

\item{pth}{p-value threshold}
}
\value{
igraph object
}
\description{
getPCorMxNetwork
}
