% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getPthwy}
\alias{getPthwy}
\title{getPthwy
for heatmapPathway app}
\usage{
getPthwy(gseaRes)
}
\arguments{
\item{gseaRes}{global variable gseaRes}
}
\value{
Pthwy
}
\description{
getPthwy
for heatmapPathway app
}
